Update section1.tsx
I'm making use of *createRef()* method to create a reference to the video element, and then using the *play()* method to start the video when the *onCanPlay* event is fired. This allows the video to start playing automatically, while also addressing the limitations of the *autoplay* attribute on Safari.

Also, I added a new method *handleEnded* that will be called when the video ends, it will reset the *currentTime* to 0 and call the play method again so the video can loop.